fix: refactor mock_dataone_formats
