+++ gwvolman/tests/__init__.py
@@ -1,5 +1,7 @@
 # A number of common/shared structures for testing
 import copy
+import httmock
+import os
 
 
 TALE = {
@@ -338,7 +340,7 @@ def mock_gc_get(path, parameters=None):
         return copy.deepcopy(TALE_NO_DESC)
     elif path.startswith("/tale") and path.endswith("/manifest"):
         assert "expandFolders" in parameters
-        assert parameters["expandFolders"] == True
+        assert parameters["expandFolders"] is True
         return copy.deepcopy(MANIFEST)
     elif path in "/tale/1cfd57fca18691e5d1feeda6":
         return copy.deepcopy(PUBLISHED_TALE)
@@ -356,3 +358,24 @@ def mock_gc_get(path, parameters=None):
         return tale
     else:
         raise RuntimeError
+
+
+@httmock.urlmatch(
+    scheme="https",
+    netloc="^cn-stage-2.test.dataone.org$",
+    path="^/cn/v2/formats$",
+    method="GET",
+)
+def mock_dataone_formats(url, request):
+    resp_fname = os.path.join(os.path.dirname(__file__), "d1formats_resp.xml")
+    with open(resp_fname, "r") as fp:
+        response = fp.read()
+    return httmock.response(
+        status_code=200,
+        content=response,
+        headers={"Connection": "Close", "Content-Type": "text/xml"},
+        reason=None,
+        elapsed=5,
+        request=request,
+        stream=False,
+    )
